machine-learning/requirements.txt
aiocache==0.12.2 
albumentations==1.3.1 
anyio==4.2.0 
certifi==2023.11.17 
charset-normalizer==3.3.2 
click==8.1.7 
colorama==0.4.6
contourpy==1.2.0 
cycler==0.12.1 
cython==3.0.8 
easydict==1.11 
exceptiongroup==1.2.0
fastapi-slim==0.111.0 
filelock==3.13.1 
fonttools==4.47.2 
fsspec==2023.12.2 
ftfy==6.2.0 
gunicorn==22.0.0 
h11==0.14.0 
httptools==0.6.1 
huggingface-hub==0.23.4 
idna==3.6 
imageio==2.33.1 
insightface==0.7.3 
joblib==1.3.2 
kiwisolver==1.4.5 
lazy-loader==0.3 
markdown-it-py==3.0.0 
matplotlib==3.8.2 
mdurl==0.1.2 
networkx==3.2.1 
numpy==1.*
# onnx==1.16.0 
opencv-python-headless==4.9.0.80 
orjson==3.10.3 
packaging==23.2 
pillow==10.3.0 
prettytable==3.9.0 
protobuf==4.25.2 
# pydantic==1.10.*
# pydantic_settings
pygments==2.17.2 
pyparsing==3.1.1 
python-dateutil==2.8.2 
python-dotenv==1.0.0 
python-multipart==0.0.9 
pyyaml
qudida==0.0.4 
requests==2.31.0 
rich==13.7.1 
scikit-image==0.22.0 
scikit-learn==1.3.2 
scipy==1.11.4 
setuptools==68.2.2 
six==1.16.0 
sniffio==1.3.0 
starlette==0.37.2 
threadpoolctl==3.2.0 
tifffile==2023.12.9 
tokenizers==0.19.1 
tqdm==4.66.1 
typing-extensions==4.9.0 
urllib3==2.1.0 
uvicorn
uvloop==0.19.0
watchfiles==0.21.0 
wcwidth==0.2.13 
websockets==12.0 
# onnxruntime==1.17.3
# onnxruntime-gpu==1.17.1
# onnxruntime-openvino==1.17.1
onnxruntime-gpu --extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-12/pypi/simple/
# onnxruntime-gpu
pyzmq==25.1.2
paho-mqtt==1.6.*
pydantic==2.3.*
pydantic-settings==2.2.1
chromadb==0.5.0
pandas==2.1.4
# torch==2.3.0
# torchvision>=0.18.0
# libcublas==11.10.3.66=0
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cublas-cu12==12.1.3.1